Test Dummy implemented, Arm and Player inherited classes for controlling
To allow for multiple Player and Arm objects in a scene, but only one set of controllable ones, Player_Controlled and Arm_Controlled were created. The classes inherit Player and Arm, respectively, and provides input functionality. With this change, a test dummy, Player_Sandbag, was created. It is now possible to test weapon damage with this dummy.