HUD visual updates, ammo miscount bug fix, base.Start() calls
Updated the HUD so the weapon diamonds are closer to the center, the health bar scales properly with screen size, and the ammo count is more compact. There was a bug with the HUD's ammo count being updated by any player within the scene; this was remedied to make only the controlled player update it. Player_Controlled and Arm_Controlled now call base.Start() in their Start functions so that they don't completely override the base classes' Start functions.