Networked projectiles
Launchers now function across the network. Projectiles get spawned and initialized across all clients. Explosions are conducted client-side to mitigate server load, resulting in a slight position/explosion time offset from the server. This does not affect health, as that is stored server-side, but it *might* affect knockback. Currently deeming this as a necessary evil.