Last attack kills work correctly-er
The last commit message was a lie. This commit features correctly working kill tracking for the last person to damage a player when that player dies. The only way to get a Death without giving another player a Kill is to have taken no damage from other players before dying; this is intentional so players cannot kill themselves to rob others of a kill.