Damage and kill take WeaponType into account
To connect with the achievement system, all damage and kill functions pass along the WeaponType that caused the kill. In the event of "last attack" kills, the type of weapon used in that attack is considered. It is unknown if achievements are working with this change, but it does not result in any conflicts with the networking code.