Separating networked classes from base classes, starting local multi
Different classes exist for networked objects, which contain a field for the base version that allows for easy script editing. It doesn't quite work yet; the gamemanager is not working and last attacker ID has some bugs. Individual controllers can control different characters in J_Smash_Plat. A new UI is there too. This is the beginning of getting local multiplayer on a single machine.