Local and networked multiplayer works well, AI needs re-implement
Local and networked multiplayer games are possible and work with minimal issues. There are some bugs that were introduced for networked matches exclusively, but they are not game-breaking and as such should be considered a stretch goal. See the stretch goal list on Trello for more on that. Due to the changes, training matches need to be implemented differently so the UI and GameManager can pick up on AI players correctly. A new folder, FinalScenes, has been created so that working scenes can be separated from development scenes.