Weapon switching changes, AK model implemented
Weapon switching now instantiates prefabs and activates/deactivates them as opposed to changing the mesh and material. This is because meshes cannot be altered to fit within the constraints of the arm, but prefabs can. This results in a glitch with online multiplayer not showing the correct weapons, but this may be fixed by adding the prefabs to the spawnable prefabs in Network Manager. The AK model has been added with four variations to test weapon switching.