Training Mode now works locally, up to three players can join
Up to three players can play Training Mode with an AI that will switch targets every few seconds. The Prefabs/Player_AI prefab is the new AI prefab that should be edited from now on. Various other changes have been made to stages, as well as a change to indicators where the color will change depending on which player is being indicated.