Altered particles, player model relates to ID, model rotates on force
The vacuum, blower, and casing particles now draw to world space so that they don't move/rotate relative to the player. The color palette of the player model now reflects the player ID. Player model will rotate in the direction an enacted force comes from, giving the illusion of "facing" enemies and locations. This will only happen on significantly large forces.